+++ zerver/lib/exceptions.py
